Integer not Int64

+++ src/moving.jl
@@ -8,7 +8,7 @@ end
 
 ######### DataArray version
 
-function moving(df::DataFrame, col::String, f::Function, n::Int64)
+function moving(df::DataFrame, col::String, f::Function, n::Integer)
   with(df, quote
        $mvg($df[$col], $f, $n)
        end)
@@ -16,12 +16,12 @@ end
 
 ############ DataFrames bang version 
 
-function mvg(x::DataArray,f::Function,n::Int64)
+function mvg(x::DataArray,f::Function,n::Integer)
   foo = [f(x[i:i+(n-1)]) for i=1:length(x)-(n-1)]
   bar = [nas(DataVector[float(n)], n-1) ; float(foo)]
 end
 
-function moving!(df::DataFrame, col::String, f::Function, n::Int64)
+function moving!(df::DataFrame, col::String, f::Function, n::Integer)
   new_col = strcat(string(f), "_", string(n))
   within!(df, quote
           $new_col  = $mvg($df[$col], $f, $n)
